revert moving msg types
