+++ framework/packages/standards/dex/src/msg.rs
@@ -1,11 +1,14 @@
 #![warn(missing_docs)]
 //! # Dex Adapter API
-use crate::contract::DexAdapter;
-use abstract_core::objects::{AnsAsset, AssetEntry};
+use abstract_core::{
+    adapter,
+    objects::{AnsAsset, AssetEntry},
+};
 use cosmwasm_schema::QueryResponses;
 use cosmwasm_std::{Decimal, Uint128};
 // re-export response types
-pub use abstract_dex_adapter_traits::types::*;
+use abstract_core::objects::DexAssetPairing;
+use cosmwasm_std::CosmosMsg;
 
 pub type DexName = String;
 pub type OfferAsset = AnsAsset;
@@ -13,7 +16,35 @@ pub type AskAsset = AnsAsset;
 
 pub const IBC_DEX_ID: u32 = 11335;
 
-abstract_adapter::adapter_msg_types!(DexAdapter, DexExecuteMsg, DexQueryMsg);
+pub type ExecuteMsg = adapter::ExecuteMsg<DexExecuteMsg>;
+pub type InstantiateMsg = adapter::InstantiateMsg<DexInstantiateMsg>;
+pub type QueryMsg = adapter::QueryMsg<DexQueryMsg>;
+
+impl adapter::AdapterExecuteMsg for DexExecuteMsg {}
+impl adapter::AdapterQueryMsg for DexQueryMsg {}
+
+/// Response for simulating a swap.
+#[cosmwasm_schema::cw_serde]
+pub struct SimulateSwapResponse {
+    /// The pool on which the swap was simulated
+    pub pool: DexAssetPairing,
+    /// Amount you would receive when performing the swap.
+    pub return_amount: Uint128,
+    /// Spread in ask_asset for this swap
+    pub spread_amount: Uint128,
+    // LP/protocol fees could be withheld from either input or output so commission asset must be included.
+    /// Commission charged for the swap
+    pub commission: (AssetEntry, Uint128),
+    /// Adapter fee charged for the swap (paid in offer asset)
+    pub usage_fee: Uint128,
+}
+
+/// Response from GenerateMsgs
+#[cosmwasm_schema::cw_serde]
+pub struct GenerateMessagesResponse {
+    /// Messages generated for dex action
+    pub messages: Vec<CosmosMsg>,
+}
 
 #[cosmwasm_schema::cw_serde]
 pub struct DexInstantiateMsg {
