+++ framework/packages/standards/staking/src/command.rs
@@ -1,4 +1,4 @@
-use crate::types::{RewardTokensResponse, StakeResponse, StakingInfoResponse, UnbondingResponse};
+use crate::msg::{RewardTokensResponse, StakeResponse, StakingInfoResponse, UnbondingResponse};
 use crate::{CwStakingError, Identify};
 use abstract_sdk::core::objects::{AssetEntry, ContractEntry};
 use abstract_sdk::feature_objects::{AnsHost, VersionControlContract};
