+++ framework/packages/standards/staking/src/lib.rs
@@ -1,6 +1,6 @@
 mod command;
 pub mod error;
-pub mod types;
+pub mod msg;
 // Export interface for use in SDK modules
 pub use command::CwStakingCommand;
 pub use error::CwStakingError;
