+++ framework/packages/standards/staking/src/msg.rs
@@ -2,12 +2,13 @@
 //! # Staking Adapter
 //!
 //! `abstract::cw-staking`
-use crate::contract::CwStakingAdapter;
-use abstract_core::objects::{AnsAsset, AssetEntry};
+use abstract_core::{
+    adapter,
+    objects::{AnsAsset, AssetEntry},
+};
 use cosmwasm_schema::QueryResponses;
+use cosmwasm_std::Empty;
 use cw_utils::Duration;
-// Re-export response types
-pub use abstract_staking_adapter_traits::types::*;
 
 /// Name of the staking provider, used by the ANS.
 pub type ProviderName = String;
@@ -15,7 +16,13 @@ pub type ProviderName = String;
 /// The callback id for staking over ibc
 pub const IBC_STAKING_PROVIDER_ID: u32 = 22335;
 
-abstract_adapter::adapter_msg_types!(CwStakingAdapter, StakingExecuteMsg, StakingQueryMsg);
+pub type ExecuteMsg = adapter::ExecuteMsg<StakingExecuteMsg>;
+pub type InstantiateMsg = adapter::InstantiateMsg<Empty>;
+pub type QueryMsg = adapter::QueryMsg<StakingQueryMsg>;
+
+impl adapter::AdapterExecuteMsg for StakingExecuteMsg {}
+
+impl adapter::AdapterQueryMsg for StakingQueryMsg {}
 
 /// A request message that's sent to this staking adapter
 #[cosmwasm_schema::cw_serde]
@@ -101,4 +108,95 @@ pub enum StakingQueryMsg {
         /// The staking token to query
         staking_token: AssetEntry,
     },
+}
+
+use cosmwasm_std::{Addr, Uint128};
+use cw_asset::AssetInfo;
+use cw_utils::Expiration;
+
+/// Possible staking targets to support staking on cosmwasm contract or cosmos Lockup module
+#[cosmwasm_schema::cw_serde]
+#[non_exhaustive]
+pub enum StakingTarget {
+    /// Address of the staking contract (Cosmwasm)
+    Contract(Addr),
+    /// Pool id of the staking contract (Osmosis)
+    Id(u64),
+}
+
+impl StakingTarget {
+    /// Extract contract address
+    pub fn expect_contract(self) -> abstract_core::AbstractResult<Addr> {
+        match self {
+            StakingTarget::Contract(addr) => Ok(addr),
+            _ => Err(abstract_core::AbstractError::Assert(
+                "Staking target is not a contract address.".into(),
+            )),
+        }
+    }
+
+    /// Extract pool id
+    pub fn expect_id(self) -> abstract_core::AbstractResult<u64> {
+        match self {
+            StakingTarget::Id(id) => Ok(id),
+            _ => Err(abstract_core::AbstractError::Assert(
+                "Staking target is not an pool ID.".into(),
+            )),
+        }
+    }
+}
+
+impl From<u64> for StakingTarget {
+    fn from(value: u64) -> Self {
+        Self::Id(value)
+    }
+}
+
+impl From<Addr> for StakingTarget {
+    fn from(value: Addr) -> Self {
+        Self::Contract(value)
+    }
+}
+
+/// Response for the staking_info query
+#[cosmwasm_schema::cw_serde]
+pub struct StakingInfoResponse {
+    /// Contract or pool id to stake to
+    pub staking_target: StakingTarget,
+    /// Staking token
+    pub staking_token: AssetInfo,
+    /// Different supported unbonding periods. None if no unbonding is supported.
+    pub unbonding_periods: Option<Vec<Duration>>,
+    /// Max number of claims. None if no limit.
+    pub max_claims: Option<u32>,
+}
+
+/// Response for the staked query
+#[cosmwasm_schema::cw_serde]
+pub struct StakeResponse {
+    /// Amount of staked tokens
+    pub amount: Uint128,
+}
+
+/// Response for the rewards query
+#[cosmwasm_schema::cw_serde]
+pub struct RewardTokensResponse {
+    /// List of reward tokens
+    pub tokens: Vec<AssetInfo>,
+}
+
+/// Response for the unbonding query
+#[cosmwasm_schema::cw_serde]
+pub struct UnbondingResponse {
+    /// List of unbonding entries
+    pub claims: Vec<Claim>,
+}
+
+/// A claim for a given amount of tokens that are unbonding.
+#[cosmwasm_schema::cw_serde]
+pub struct Claim {
+    /// Amount of tokens that are unbonding
+    pub amount: Uint128,
+    /// When the tokens can be claimed
+    pub claimable_at: Expiration,
 }
