+++ integrations/astroport/packages/abstract-adapter/Cargo.toml
@@ -21,7 +21,7 @@ full_integration = [
 [dependencies]
 cosmwasm-std = { version = "1.1" }
 abstract-staking-adapter-traits = { version = "0.18.0", path = "../../../../framework/packages/standards/staking" }
-abstract-dex-adapter-traits = { version = "0.18.0", path = "../../../../framework/packages/standards/dex" }
+abstract-dex-standard = { version = "0.18.0", path = "../../../../framework/packages/standards/dex" }
 abstract-sdk = { version = "0.18.0", path = "../../../../framework/packages/abstract-sdk" }
 
 
@@ -34,7 +34,7 @@ astroport = { path = "../astroport/", optional = true }
 lazy_static = "1.4.0"
 
 [dev-dependencies]
-abstract-dex-adapter-traits = { version = "0.18.0", features = [
+abstract-dex-standard = { version = "0.18.0", features = [
   "testing",
 ], path = "../../../../framework/packages/standards/dex" }
 cw-orch = { version = "0.16.1", features = ["daemon"] }
