+++ integrations/kujira/packages/abstract-adapter/Cargo.toml
@@ -15,7 +15,7 @@ full_integration = ["dep:kujira", "dep:cw20", "dep:cw-asset", "dep:cw-utils"]
 cosmwasm-std = { version = "1.1.2", features = ["stargate"] }
 abstract-core = { version = "0.18.0", path = "../../../../framework/packages/abstract-core" }
 abstract-staking-adapter-traits = { version = "0.18.0", path = "../../../../framework/packages/standards/staking" }
-abstract-dex-adapter-traits = { version = "0.18.0", path = "../../../../framework/packages/standards/dex" }
+abstract-dex-standard = { version = "0.18.0", path = "../../../../framework/packages/standards/dex" }
 kujira = { version = "0.8.2", optional = true }
 
 #Abstract dependencies 
@@ -25,7 +25,7 @@ cw-asset = { version = "3.0.0", optional = true }
 cw-utils = { version = "1.0.1", optional = true }
 
 [dev-dependencies]
-abstract-dex-adapter-traits = { version = "0.18.0", path = "../../../../framework/packages/standards/dex", features = [
+abstract-dex-standard = { version = "0.18.0", path = "../../../../framework/packages/standards/dex", features = [
   "testing",
 ] }
 cw-orch = { version = "0.16.1", features = ["daemon"] }
