+++ integrations/osmosis-adapter/Cargo.toml
@@ -20,7 +20,7 @@ full_integration = [
 cosmwasm-std = { version = "1.1.2", features = ["stargate"] }
 abstract-core = { version = "0.18.0", path = "../../framework/packages/abstract-core" }
 abstract-staking-adapter-traits = { version = "0.18.0", path = "../../framework/packages/standards/staking" }
-abstract-dex-adapter-traits = { version = "0.18.0", path = "../../framework/packages/standards/dex" }
+abstract-dex-standard = { version = "0.18.0", path = "../../framework/packages/standards/dex" }
 
 #Abstract dependencies 
 cw20 = { version = "0.15", optional = true }
