+++ modules/contracts/adapters/cw-staking/src/adapter.rs
@@ -1,4 +1,4 @@
-use crate::msg::StakingAction;
+use abstract_staking_adapter_traits::msg::StakingAction;
 use abstract_staking_adapter_traits::CwStakingError;
 use cosmwasm_std::{DepsMut, Env, MessageInfo, SubMsg};
