+++ modules/contracts/adapters/cw-staking/src/handlers/execute.rs
@@ -1,11 +1,13 @@
 use crate::adapter::CwStakingAdapter;
 use crate::contract::{CwStakingAdapter as CwStakingContract, StakingResult};
-use crate::msg::{ProviderName, StakingAction, StakingExecuteMsg, IBC_STAKING_PROVIDER_ID};
 use crate::resolver::{self, is_over_ibc};
 use abstract_sdk::core::ibc_client::CallbackInfo;
 use abstract_sdk::feature_objects::AnsHost;
 use abstract_sdk::features::{AbstractNameService, AbstractResponse};
 use abstract_sdk::{IbcInterface, Resolve};
+use abstract_staking_adapter_traits::msg::{
+    ProviderName, StakingAction, StakingExecuteMsg, IBC_STAKING_PROVIDER_ID,
+};
 use cosmwasm_std::{to_binary, Coin, Deps, DepsMut, Env, MessageInfo, Response};
 
 const ACTION_RETRIES: u8 = 3;
