+++ modules/contracts/adapters/dex/src/contract.rs
@@ -1,9 +1,9 @@
 use crate::handlers;
 use crate::EXCHANGE;
 
-use crate::msg::{DexExecuteMsg, DexInstantiateMsg, DexQueryMsg};
 use abstract_adapter::{export_endpoints, AdapterContract};
-use abstract_dex_adapter_traits::DexError;
+use abstract_dex_standard::msg::{DexExecuteMsg, DexInstantiateMsg, DexQueryMsg};
+use abstract_dex_standard::DexError;
 use cosmwasm_std::Response;
 
 pub const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
