+++ modules/contracts/adapters/dex/src/exchanges/exchange_resolver.rs
@@ -1,7 +1,7 @@
 use abstract_adapter_utils::identity::decompose_platform_name;
 use abstract_adapter_utils::identity::is_available_on;
 use abstract_adapter_utils::identity::is_current_chain;
-use abstract_dex_adapter_traits::{DexCommand, DexError, Identify};
+use abstract_dex_standard::{DexCommand, DexError, Identify};
 use cosmwasm_std::Env;
 
 /// Any exchange should be identified by the adapter
