+++ modules/contracts/adapters/dex/src/exchanges/terraswap.rs
@@ -1,4 +1,4 @@
-use abstract_dex_adapter_traits::Identify;
+use abstract_dex_standard::Identify;
 
 pub const TERRASWAP: &str = "terraswap";
 
@@ -17,8 +17,8 @@ impl Identify for Terraswap {
 #[cfg(feature = "terra")]
 use ::{
     abstract_core::objects::PoolAddress,
-    abstract_dex_adapter_traits::{coins_in_assets, cw_approve_msgs},
-    abstract_dex_adapter_traits::{DexCommand, DexError, Fee, FeeOnInput, Return, Spread},
+    abstract_dex_standard::{coins_in_assets, cw_approve_msgs},
+    abstract_dex_standard::{DexCommand, DexError, Fee, FeeOnInput, Return, Spread},
     abstract_sdk::cw_helpers::wasm_smart_query,
     cosmwasm_std::{to_binary, wasm_execute, CosmosMsg, Decimal, Deps},
     cw20::Cw20ExecuteMsg,
