+++ modules/contracts/adapters/dex/src/handlers/execute.rs
@@ -2,11 +2,11 @@ use crate::handlers::execute::exchange_resolver::is_over_ibc;
 
 use crate::contract::{DexAdapter, DexResult};
 use crate::exchanges::exchange_resolver;
-use crate::msg::{DexAction, DexExecuteMsg, DexName, IBC_DEX_ID};
 use crate::state::SWAP_FEE;
 use abstract_core::objects::account::AccountTrace;
 use abstract_core::objects::chain_name::ChainName;
-use abstract_dex_adapter_traits::DexError;
+use abstract_dex_standard::msg::{DexAction, DexExecuteMsg, DexName, IBC_DEX_ID};
+use abstract_dex_standard::DexError;
 
 use abstract_core::ibc_client::CallbackInfo;
 use abstract_core::objects::ans_host::AnsHost;
