+++ modules/contracts/apps/dca/src/handlers/execute.rs
@@ -35,9 +35,7 @@ fn create_convert_task_internal(
         actions: vec![CronCatAction {
             msg: wasm_execute(
                 env.contract.address,
-                &ExecuteMsg::from(DCAExecuteMsg::Convert {
-                    dca_id: dca_id.into(),
-                }),
+                &ExecuteMsg::from(DCAExecuteMsg::Convert { dca_id: dca_id }),
                 vec![],
             )?
             .into(),
